Generify Dropdown model
