+++ src/main/java/madstax/model/util/DropdownModel.java
@@ -1,0 +1,29 @@
+package madstax.model.util;
+
+import javax.swing.*;
+
+public class DropdownModel<E> extends DefaultComboBoxModel<E> {
+
+    private boolean introSelectionAllowed;
+
+    public DropdownModel(E[] items) {
+        super(items);
+        this.introSelectionAllowed = true;
+    }
+
+    @Override
+    public void setSelectedItem(Object anObject) {
+        if (anObject == null && introSelectionAllowed) {
+            String message = "Select an option";
+            super.setSelectedItem(message);
+        } else {
+            introSelectionAllowed = false;
+            super.setSelectedItem(anObject);
+        }
+    }
+
+    @Override
+    public E getSelectedItem() {
+        return (E) super.getSelectedItem();
+    }
+}
